powerpc/book3e: IBM Blue Gene/Q Quad Processing eXtention (QPX)
This enables kernel support for the QPX extention and is intended for
processors that support it, usually an IBM Blue Gene processor.
Turning it on does not effect other processors but it does add code
and will quadruple the per thread save and restore area for the FPU
(hense the name).  If you have enabled VSX it will only double the
space.

Signed-off-by: Jimi Xenidis <jimix@pobox.com>