powerpc/book3e: IBM Blue Gene/Q QPX: Increase FPU storage size for QPX.
Increase the per-thread storage that the Quad Processing eXention
register set requires.  Like VSX the QPX sues the same register sapce
for FPU and vector registers so we continue to use the FPU sotrage
space, just increase its size.

The space is 4 times that of the normal FPU, which is why the unit has
been nicknamed the "Quad-Hummer".  The register profile is:
  - 4-wide double precision FPU SIMD
  - 2-wide complex SIMD
  - 4R/2W register file (32x256 bits per thread)

Signed-off-by: Jimi Xenidis <jimix@pobox.com>