powerpc/book3e: IBM Blue Gene/Q QPX: Save and restore the QPX registers using the FPU path.
Since the QPX (like the VSX) unified the FPU and the vector unit, we
can save the register set of both units by saving all vector unit
registers.  Since VSX works in a similar fashion _and_ it is
desireable to have a single binary be able to use the QPX _or_ the VSX
we try to share as much code as possible.  However, the addition of
the third option made the code unreadable so the save and restore
segments of code are layed out seperately and branched to according to
the CPU feature.  This streategy should be (a little) easier to read
but also reduced the about of feature NOPing at runtime.

Signed-off-by: Jimi Xenidis <jimix@pobox.com>